+++ history.rst
@@ -16,8 +16,12 @@ Version 0.7.2
 
 - Better checking if empty json string.
 
-- Workaround for false "Content-Types": If the request is a POST-request,
-  the body will not read by cherrypy.
+- Cherrypy:
+
+  - Workaround for false "Content-Types": If the request is a POST-request,
+    the body will not read by cherrypy.
+
+  - Bug in GET-requests fixed: GET-requests are possible now.
 
 
 =============